Changed prompt type to enum in params
