+++ IdentityCore/src/util/MSIDPromptType_Internal.h
@@ -21,32 +21,8 @@
 // OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 // THE SOFTWARE.
 
-#import "MSIDTelemetryBaseEvent.h"
+#import <Foundation/Foundation.h>
 #import "MSIDConstants.h"
 
-@class MSIDAccount;
-@class MSIDAuthority;
-
-@interface MSIDTelemetryAPIEvent : MSIDTelemetryBaseEvent
-
-- (void)setCorrelationId:(NSUUID *)correlationId;
-- (void)setExtendedExpiresOnSetting:(NSString *)extendedExpiresOnSetting;
-- (void)setUserId:(NSString *)userId;
-- (void)setClientId:(NSString *)clientId;
-- (void)setIsExtendedLifeTimeToken:(NSString *)isExtendedLifeToken;
-- (void)setErrorDomain:(NSString *)errorDomain;
-
-- (void)setApiId:(NSString *)apiId;
-
-- (void)setWebviewType:(NSString *)webviewType;
-
-- (void)setLoginHint:(NSString *)loginHint;
-- (void)setErrorCode:(NSUInteger)errorCode;
-- (void)setPromptType:(MSIDPromptType)promptType;
-
-- (void)setIsSuccessfulStatus:(NSString *)successStatus;
-- (void)setResultStatus:(NSString *)resultStatus;
-- (void)setUserInformation:(MSIDAccount *)account;
-- (void)setOauthErrorCode:(NSString *)oauthErrorCode;
-
-@end
+extern NSString * MSIDPromptParamFromType(MSIDPromptType type);
+extern MSIDPromptType MSIDPromptTypeFromString(NSString *promptTypeString);
